+++ db/migrations/20200505135053_register_legacy_extensions.php
@@ -1,0 +1,36 @@
+<?php
+
+/*
+ * This file is part of the Eventum (Issue Tracking System) package.
+ *
+ * @copyright (c) Eventum Team
+ * @license GNU General Public License, version 2 or later (GPL-2+)
+ *
+ * For the full copyright and license information,
+ * please see the COPYING and AUTHORS files
+ * that were distributed with this source code.
+ */
+
+use Eventum\Db\AbstractMigration;
+use Eventum\Extension\Legacy;
+use Eventum\Extension\RegisterExtension;
+
+class RegisterLegacyExtensions extends AbstractMigration
+{
+    private const EXTENSIONS = [
+        Legacy\CustomerLegacyExtension::class,
+        Legacy\WorkflowLegacyExtension::class,
+    ];
+
+    public function up(): void
+    {
+        $register = new RegisterExtension();
+        $register->register(...self::EXTENSIONS);
+    }
+
+    public function down(): void
+    {
+        $register = new RegisterExtension();
+        $register->unregister(...self::EXTENSIONS);
+    }
+}