+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Upgrading to 3.8.x versions requires that you upgrade to latest 3.5.x version fi
 - Catch and log fatal errors if XhguiProfile fails to initialize, #822
 - Migrate `SPHINX_SEARCHD_HOST` constant to `setup.php`, #824
 - Add Pimple based service container, #826
+- Add `ACCESS_ISSUE` event for checking issue access, #830
 
 [3.8.11]: https://github.com/eventum/eventum/compare/v3.8.10...master